pyproject.toml
[tool.poetry]
name = "f1-poetry"
version = "0.1.0"
description = "Final Poetry file"
authors = ["Aniket <144729623+aniket-afk@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.5"
streamlit = "^1.40.2"
pinecone-client = "^5.0.1"
openai = "^1.55.1"
apache-airflow = "^2.10.3"
uvicorn = "^0.32.1"
snowflake-connector-python = "^3.12.3"
python-dotenv = "^1.0.1"
boto3 = "^1.35.71"
selenium = "^4.27.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
docling = "^2.8.3"
docling-core = "^2.7.0"
langchain-text-splitters = "^0.3.2"
langchain-huggingface = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
